% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{clean_data}
\alias{clean_data}
\title{clean_data
Helper function that formats the output of the query_api function into a tibble with appropriate column types}
\usage{
clean_data(query_api_output)
}
\arguments{
\item{query_api_output}{the output provided by the \code{query_api} function.}
}
\description{
clean_data
Helper function that formats the output of the query_api function into a tibble with appropriate column types
}
