% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_key}
\alias{get_key}
\title{Retrieves the API key}
\usage{
get_key()
}
\value{
The user's API key, if set as an environment variable. If key is not set, returns an error message indicating that the user can either set
the environment variable \code{CN_API} or pass the API key as an argument to the function.
}
\description{
\code{get_key} stops execution and returns an error message if the environment variable \code{CN_API} is not set to the user's API key. If the environment
variable is not set, the function returns an error message.
}
\examples{
get_key()
}
\references{
Note that this function is based on, and repeats much of the code in
the R API package best practices, available at \url{https://httr.r-lib.org/articles/api-packages.html#overall-design-1}.
}
